Removed (Long)

+++ c_version/kaprekarUtils.c
@@ -29,7 +29,7 @@ int isKaprekar(int n) {
   }
 
   int i;
-  long square = n * (long) n;
+  long square = n * n;
   int numDigits = (int) log10(n) + 1;
   long modulus = 0;
   long first, second;
